+++ lib/supply_chain.dart
@@ -23,6 +23,6 @@ export 'src/producer.dart';
 export 'src/scm.dart';
 export 'src/scope_blue_print.dart';
 export 'src/scope.dart';
-export 'src/sub_scope_manager_blue_print.dart';
-export 'src/sub_scope_manager.dart';
+export 'src/scope_blue_print_factory.dart';
+export 'src/scope_factory.dart';
 export 'src/tools.dart';